requirements.txt
aiogram==3.17.0
openpyxl==3.1.5
python-dotenv==1.0.1
aiosqlite==0.21.0
